Use Preconditions from same repo.
This is a temporary measure to allow us to use Preconditions before it's in NuGet.

Unfortunately using a relative path including .. from global.json only appears to
work when dnu restore is run from the solution directory, instead of the parent, hence
the change to .travis.yml. See http://stackoverflow.com/questions/33919272 for
details of attempts.