Update StorageClient to follow the same static creation pattern as generated APIs
To create a StorageClient, there are now four options:
- Pass in an existing StorageService
- Pass in an existing initializer (less likely to be useful; drop?)
- Call the static Create method and pass in optional credentials
- Call the static CreateAsync method and pass in optional credentials

Now we *could* have a synchronous constructor which uses Task.Run to get the credentials,
basically like Create. In our other APIs we have an abstract base class and then a concrete derived
class, making the constructor route a no-no... I haven't gone for that here, although we could if
we wanted to.

We could also have an extension method of ToStorageClient on StorageService if that proved useful.