Rewrite of the snippet generation tool
This now generates two files per type: a snippet text file (which currently just
contains the extracted snippets, but could have mutations in the future) and
a markdown file with the UID mappings.

The snippets are now declared in a different way, which is more flexible and doesn't have
to conform to docfx snippet ID rules.