First part of Datastore "wrapping" improvements.
Ignore the fact that it's called DatastoreFoo - we need to work out a better name, but haven't come up with one yet.
DatastoreFooSnippets is a copy of DatastoreClientSnippets, but using DatastoreFoo everywhere.

It's definitely a big improvement over DatastoreClient, but I want to do more with query construction before presenting it
to the Datastore team.