Remove Microsoft copyright from footer
I haven't replaced it with a Google copyright notice, as it looks like none of
our other documentation has a copyright in the page.

The use of a space is deliberate here - an empty string didn't make any difference, i.e.
the original Microsoft copyright was still shown.

This fixes issue #167.