Add GAX to the generated API documentation
This is achieved by adding a git submodule under docs/external.
There are lots of reasons *not* to use git submodules, but I think in this case it's
probably the right approach - at least for now. We'll need to update the submodule to take account of
GAX changes, but it means we *can* pin it to a specific version.

The way GAX currently exists, this means we end up with quite a bit of cruft in the Google.Api
namespace as well; we might want to fix that in the future. (https://github.com/googleapis/gax-dotnet/issues/39 would do that too.)
There's a broader discussion to be had about what documentation we want where, but that can wait.