Enable resources to be specified in snippets
The content is added into the text file for the C# code that refers to the resource.
This is all a bit ugly, but works for now...
In the same change, I've moved away from using "Sample" to determine whether or not to use the
snippet ID as a member; instead, for snippets we just add the snippet ID as a member to match.