Manually-written change to use channel pool scoping.
Note that the declaration of s_channelPool must be moved to after
the declaration of DefaultScopes as otherwise DefaultScopes would be
null at the point of pool creation.

This change should end up in the code generator, but for now this
simple change will unblock us.

All versions of packages from gax-dotnet upgraded to CI00074 for
simplicity.