Handle repeated fields and records in ListRows
Although repeated fields and records aren't exposed in queries, you
can access them via listing the rows in a table, and we weren't
handling them properly.

We now expose them as arrays (for repeated fields) and dictionaries
of type Dictionary<string, object> (for records). It's potentially
not quite ideal, but it's pretty reasonable for a relatively rare use
case. Feedback would be useful from actual usage. (It's likely I'll
need this when making performance data available in Noda Time, so
hopefully I'll discover the usability then.)

Fixes #324 and #325.