Modify project.json to include beta01 for beta APIs
This means that our CI builds will become 1.0.0-beta01-CI00100 etc, which means
beta02 is later than any of those, etc - which is what we want.

Note that the version in project.json should refer to the "most recently released" version;
we can't accidentally push that a second time (as nuget will stop us from doing so) and it
means that CI build subscribers will see "the right thing". There should be a change to
update to the next version (e.g. beta02) at the same time as that version is released.