Completely change our docs layout (#348)
* Completely change our docs layout

We now have a directory per API, as well as a top-level, API-neutral
directory (for articles such as page-streaming). Each API-specific
directory contains API-specific articles and only the relevant
dependencies in its API reference. This makes it much easier to
find things when dealing with a single API, as well as making it
feasible to include those dependencies.

As well as reorganizing things, there are new projects in this
commit:

- tools\Google.Cloud.Tools.Common for common code between all tools
- tools\Google.Cloud.Tools.GenerateDocfxSources to generate
  docfx.json and toc.yml automatically
- tools\Google.Cloud.Tools.ValidateLinks to validate relative hrefs
  in the resulting site when assembled, as docfx can't do it as we
  go
- tools\Google.Cloud.Docs.Snippets as a new home for top-level
  snippets (e.g. page-streaming, resource names)

Still to do:

- Integration of generated snippets
- Generation of index.md for APIs which have no handwritten code
- Versioning considerations (current code has a single version)
- Inclusion of dependencies for REST APIs (e.g. Google.Apis.Storage.v1)

* Restore tools and apis together