Rename the class in all generated snippets to have a "Generated" prefix
(This will be done in the generator itself too, of course.)

This avoids the generated snippet class from interfering with the manually-written one.
The filename staying the same means it's still easy to find in Visual Studio, and the
generated code will implicitly be a "child" (in VS) of any manual snippets.

The snippet code consumer needs to take account of all of this too, of course.