Expose new gRPC methods to create IAM clients.
This is ugly at the moment, modifying generated code. We have filed
https://github.com/grpc/grpc/issues/8506 to allow gRPC to generate
partial classes, at which point the change can be in a separate file.

The purpose of these methods is to allow for toolkit code to
create a gRPC client appropriately to expose IAM methods on both
PublisherClient and SubscriberClient.