First pass at parameterized queries (#497)
* First pass at parameterized queries

We don't support Strct parameters yet, but I think we can probably
live without that for v1.0 - and even if we decide later that we
need to add them, this is still worth merging before that.

* Simplify rules for parameter type inference for System.DateTime values

Rather than inferring the type based on the Kind of the
value, it's simpler just to say that System.DateTime is always inferred
as a BigqueryParameterType.DateTime parameter. If the caller wants
to use a UTC DateTime as a BigqueryParameterType.Timestamp, they can
easily specify the type explicitly.