Edit the Datastore code-generated snippets
The C# protoc plugin removes a prefix of the enum name from enum
values. We need to do the same in the toolkit generator.

(As an aside, this value is probably never valid - arguably when
providing a sample with an enum value, we should use the first
non-zero value... that's a bigger debate though.)