Seal various classes and turn public fields (!) into properties
Added rudimentary code health tests - just for the REST
libraries at the moment. These could of course be implemented
in any number of other ways, but this felt like the simplest
way to get started, and it spotted various issues...

Fixes #160 along the way. (Not for the gRPC clients, but they
don't have all the options classes anyway, which was my main concern.)