Sketch of the hash validating code.
This has a copy of the relevant REST support library code (and one
generated class) for ease of testing - we'd need to make changes to
the REST support library, then use those from this package. (So only
the Hash* code would actually be new.)

I've moved away from Crc32c extending HashAlgorithm as the
netstandard version doesn't expose enough members to be useful.

This code appears to work for download validation, but doesn't help
for upload - nothing appears to check it. I'm tempted to make the
changes to the upload REST code anyway so that it's easy to
implement later on.