Trace Instrumentation Library (#541)
* Move the Cloud Error Reporting Excpetion Logger into its own directory.

* Initial check in of CloudTrace Instrumentation Library

* Mark classes as internal.

* Misc clean up

* Move Cloud Error Reporting to its own directory

* Unit tests for the CloudErrorReportingExceptionLogger.

* Remove dotnet install workaround - fixed now

* Metadata emulator wait for change support (#526)

* Remove 'GetClientMock' function adn use mock 'VerifyAll' call.

* Changes from PR #541

* Fix doc typos

* Fix intermittent test failures and remove hard-coded value (#543)

* Last ETag support in MetadataClient (#544)

* First pass at docs for Google.Devtools.AspNet

This is just enough to get started, basially.
Also adjusted the example in the class itself to compile.

* Only check for file path and line number on windows as it seems to fail on mono

* Changes from PR #541

* Add ASP.NET devtools to doc root

* Update example in code to be correct and to match example in docs.

* Remove public readonly fielnds and update comment.

* Add code tags around example code.

* Add summary comments to public getters

* Fixed some docs comments with the wrong format (#553)

* Support for adding/removing metadata values in the emulator. (#552)

* Implementing default preconditions.

For an object, both the generation and metageneration are used as
preconditions. For a bucket, there's only the metageneration.

This doesn't apply a precondition when uploading new data; it
only affects metadata (UpdateObject/UpdateBucket). We may wish to
change this decision later, based on feedback.

* Simple app to delete any buckets left over from testing

These aren't easy to delete from the console due to potentially
having multi-version files.

(We should make it easier to delete them from code admittedly,
but that's covered in an issue already.)

* Remove restriction on update object/bucket

The restriction has been removed on the server side (as proved by
snippet and integration tests) so we don't need to guard against
it in the client code. Docs amended too.

This means we can close #159 as obsolete.

* Tests for MetadataClient events (#551)

* Updates from PR #541

* Move the Cloud Error Reporting Excpetion Logger into its own directory.

* Initial check in of CloudTrace Instrumentation Library

* Mark classes as internal.

* Misc clean up

* Changes from PR #541

* Changes from PR #541

* Remove public readonly fielnds and update comment.

* Add summary comments to public getters

* Updates from PR #541