Add simpler support for GqlQueryParameters
Note that this only populated a GqlQueryParameter with a Value, never a Cursor.
While we could add overloads for cursors, the parameter type would be ByteString...
which can also be converted to Value. So I'm basically betting that when you've got a
ByteString, you're more likely to want to use it as a Value parameter than as a Cursor
parameter.

Fixes #887