Update docfx filters
We used to filter Foo.Types, which is always just a "container"
class for nested protobuf types - but this appears to then *also*
remove the nested types themselves, which is counterproductive.

Without that filter, we get the class we don't want - but we get the
classes we *do* want, which is a reasonable compromise.

Fixes #982.