Change how external dependencies are handled in docs
This allows us to use a mixture of project.json and csproj projects.

The metadata for external dependencies is now built when the
"dependencies" directory is first created. (That can be recreated by
calling fetchdependencies.sh.) The result is a api subdirectory
containing all the metadata, broken down by package. The
GenerateDocfxSources code detects the dependencies, and then the
builddocs.sh script copies it into place alongside the other library
metadata, and concatenates the TOC files.

The "dependencies" directory is actually hosted in github as a
separate (orphan) branch to avoid having to recreate in every CI
AppVeyor build; it will need updating when the dependencies are
updated.

It would have been nice to be able to just refer to the external
dependency metadata from the docfx.json files, but if the files
start in different directories that ends up being tricky.