Add integration tests for requester-pays feature
These require two extra environment variables
(REQUESTER_PAYS_TEST_PROJECT and REQUESTER_PAYS_CREDENTIALS) to be
set. If they aren't set, requester-pays testing is not performed.

Note that within the fixture code, we have to specify the
UserProject option for all operations - I don't *think* we should
have to do that, but it looks like it's required right now.