Refactor BigQuery query handling
ExecuteQuery is now just a shortcut for CreateQueryJob followed by calling GetQueryResults.
- ExecuteQueryOptions is removed
- CreateQueryJobOptions renamed to just QueryOptions
- ExecuteQuery now accepts a GetQueryResultsOptions as well as a QueryOptions

BigQueryResults now always represents a *completed* query:
- GetQueryResults calls the HTTP API of GetQueryResults with a MaxResults of 0.
- Table.ListRows is used to fetch the actual data. This is the most efficient way of fetching results
- PollQueryUntilCompleted is removed, as GetQueryResults covers it

There may be multiple GetQueryResults HTTP API calls; each has a timeout of 1 minute, but the overall client-specified timeout may be greater.
This is preferable to using job polling, as GetQueryResults is effectively a "hanging get" - it will return as soon as the query has completed.

BigQueryResults.GetRows() is now redundant, as BigQueryResults implements IEnumerable<BigQueryRow>.
In this commit there's a "dummy" GetRows() method, to be removed in the next commit - this allows this main commit to be smaller,
as many tests call GetRows.