Prototype of Translation v2 (breaking change for models)
This treats models as simple strings. No validation is performed.
Everything defaults to null, meaning "don't specify any model".
If a model is specified as the default for a client, then a null
argument in a method call means "use the default" but an empty string
means "wipe out the default; don't send any model in the request."

All of this is a breaking change, hence the major version bump.
We don't know if we want to do this - this PR is intended to help
the conversation.