Remove the option to log exceptions straight to the error reporting API
The error reporting API has not gone GA yet so these libraries cannot GA with this option.

This is done by removing the EventTarget for error reporting.  This is the only externally exposed item being remove.  If we want to add back this option when the API GAs we can add back the type and add a function to create an EventTarget with it.

We still have a dependency on the Error Reporting API in the test to get the events.