Introduce a single place for tool versions to be specified
This also declares variables to run the tools, and installation
scripts. Arguably as it contains more than version numbers,
toolversions.sh should be called something else - suggestions welcome.

docfx is now installed within the scripts that use it rather than as
part of the AppVeyor script: it doesn't matter which version of
docfx a user might have on the path.

In tidying up tooling in general, I've removed coveralls.sh which is
no longer needed (as we're using codecov).

The generateapis.sh script is slightly different from all the rest
here; we can address that later.

Fixes #1324.