While individual rpc methods within spanner.data are all idempotent, the entire transaction is not. Therefore, allowing retries at the transaction level on deadline_exceeded is especially dangerous.
Also, unavailable is already retried at the rpc level -- so allowing another set of exponential backoff
retries at the transaction level is unnecessary (but not as harmful as deadline_exceeded).

The only error that should be retried at the transaction level is aborted -- which indicates a transaction deadlock.
Verified that this behavior is now fully in sync with the java client.