Implement BigQuery dataset label manipulation
This *isn't* part of BigQueryDataset at the moment, partly as it's not clear how
the current get/patch model would interact with the in-memory data. We can revisit that
if we get feature requests for it.

The quota for dataset modification operations is limited to one every two seconds, so
various tests have deliberate pauses.