Generate documentation stubs in project generator
Additionally, instead of listing every file we might have changed,
we just list files that *have* changed (or been created).

Tools solution file change is a no-op, effectively.

Generating the stubs ended up creating one for Bigtable, so I've
just modified it to say there'll be more on the way.