Modify project generator to use latest test dependencies
This includes xUnit analyzers by default. Some exceptions are thrown by the analyzers,
so we update to 0.6.1 to fix some of them, and ignore AD0001 to avoid the rest being a problem.

A couple of xUnit analyzers are misguided IMO, so I've ignored those in all projects.

At the same time, I've removed CS4014 (await required) as the generated snippets now do the right thing.