Update build procedure to help Travis
Instead of building everything in one giant solution, build one API
at a time. This is a bit less efficient, but it means we don't have
one huge build that stalls on Travis for over 10 minutes. We need to
explicitly ignore ASP.NET as that solution only builds on .NET
Desktop. Likewise we don't build the tools solution or tests on
Linux, although some projects will be pulled in anyway.

The Speech solution file change is to remove an integration test
project which doesn't actually exist. The language generator now
targets netcoreapp2.0 so it can still use WebClient but work on .NET
Core.

The BigQuery "clean test data" project is fine to build on .NET Core.