Fix release process
The argument to buildrelease.sh is now a commit hash. The script
clones the repo as normal and checks out the specified commit, then
finds all the tags for that commit, failing if there aren't any. The
set of projects to build is the set of tags, truncating each tag at
"-" (as the part after "-" is the version number in the tag).

We now only build, pack and document the specified projects.