Move to a simpler query approach
Overall intention: become less ADO-like, to avoid confusion with a
later dedicated ADO package. Currently we're in a bit of an uncanny
valley: it's deliberately ADO-like, but without actually supporting
ADO. Even without the prospect of an ADO package that has its
disadvantages, and it's more clearly "wrong" when there's an actual
ADO package.

BigQueryCommand and BigQueryParameterCollection become obsolete, as
a collection of BigQueryParameter values are provided directly on
calls to BigQueryClient. The old calls (taking either just a string
or a BigQueryCommand) are now obsolete. I've left them here for now
so that it's not a completely breaking change, and they have some
time to migrate.

The new tests are copies of the legacy tests, just using the new
overloads. All snippets have been migrated.

The "parameters" parameter is deliberately required, as I expect it
to be used in most cases, and we want to encourage query
parameterization. Additionally, if it were optional, a call to
ExecuteQuery("sql here") would be ambiguous between the old and new
overloads.

A call to ExecuteQuery("sql here", null) will now call the new
overload instead of the old one, but the effect will be the same -
no parameters, default options.