Initial cuts of FirestoreDb/CollectionReference/DocumentReference/Query
These are just enough to establish the links between the classes,
and make DocumentReference available for value conversion in a later commit.

All of these will be expanded significantly later.

Having created EqualityTester, I'll backport the Blob/GeoPoint/Timestamp/FieldPath tests to use it in a later commit too.
(Trying to backport it nicely across the commits would be very hard.)