Create RetryHandler infrastructure for BigQuery
This allows any REST service request to be marked as retriable, in which case it will be automatically retried
with exponential backoff (initially 1s, up to 32s) if the JSON response contains particular reasons.

The handler will be installed in any BigqueryService passed to a BigQueryClientImpl constructor, if it doesn't already exist.
Currently this isn't done in an atomic way, but to do so will require new functionality in the REST services support library.
(This is a "good enough" starting point.)

Code to take advantage of this functionality will be added in a subsequent commit.