Introduce WriteBatch
WriteBatch is at the heart of all mutations. Create is just the first operation - we'll also have Set and Update.
I expect WriteBatch to be thoroughly unit tested, but other integration points (soon to be added, such as
DocumentReference.Create) to be less thoroughly unit tested, with integration tests instead. (Most of those
will be convenience methods calling into WriteBatch, so the testing should stlil be good overall.)