Introduce the Update operation for documents
The precise semantics are complicated - the documentation at
https://firebase.google.com/docs/firestore/manage-data/add-data
only shows simple examples.

With this commit, the caller has to pass a Dictionary<FieldPath, object> to the Update method, which is painful.
The plan is to get the core functionality in and correct, then consider how to make it simpler to use.