Implement sentinel values for Firestore
The internal representation of sentinel values is ghastly: we use
the fact that NullValue is a single-value enum, but that protos can
store undefined enums. So the temporary Value representations of
sentinel values are highly unusual null values... but fortunately
they'll never be visible to the user, or transmitted on the wire.

Everything else about this is *reasonably* straightforward. There's
an odd situation where a field has a map value, but all the
sub-values are sentinel values; we end up transmitting an empty map,
which *may* not be what's desired. Will check that with the
Firestore team, but I don't think it should block this PR.