Fix Translation tests when using updated REST support libraries
Translation is the only service that still uses the "data" wrapper,
and the service's SerializeObject method is normally used to
serialize a *request*. Our tests use it to serialize a *response*,
so when we fixed a bug which was accidentally wrapping requests in a
"data" element, that broke these tests.

The fix is simple, and this issue should only affect users of the
Translation service who are also serializing responses with
SerializeObject. I suspect that set is basically "this code".