First pass at general testing documentation with API-specific details
This deletes the Bigtable-specific README, putting the details in
TESTING.md instead. I removed the duplication of the TEST_PROJECT
variable, but kept the part about account permissions, as I wasn't
sure whether the regular service account for a project owner was
enough. (I'll try that on Monday and slim the doc down if possible.)

I *think* these three are the only APIs that require additional
instructions for now.