Provide an escape hatch for hash validation failures
This is the first part of solving #1641. It's only a workaround, but
it's a start.

Note that there's no HashValidationMode.WhenAvailable yet; I'm not
convinced that we have a reliable signal for when we won't be able
to validate the hash, so I'd rather not add the enum value yet.
If/when we get that signal, we can add the enum value.

The aim of this PR is to unblock users until a fuller solution is
available, without delaying a GA release of Google.Cloud.Storage.V1 v2.1.0.