Only build changes in PRs
Adds a new --diff option to build.sh, which finds which APIs have
changed vs master. Only those APIs are built. If no APIs are
detected (as per this PR...), all APIs are built.

On merge, I *think* we'll find an empty diff, and so automatically
rebuild all APIs. Hard to tell until we merge this...

Nuance: if we change something like Google.Cloud.Spanner.V1, we
won't automatically rerun tests for Google.Cloud.Spanner.Data until
the merge. Basically we rely on developers to be sensible about
that, and for the merge to pick up changes.

Bonus: running build.sh --diff from a developer machine on their
branch works fine too...