Implement Spanner batched reads
This introduces two new concepts:

- The ability to get a descriptor of a SpannerTransaction (TransactionId)
  which allows you to reconnect to that transaction in another process.

- The concept of a CommandPartition which represents a Command which runs
  on a subset of data. A CommandPartition can be re-hydrated into a proper
  SpannerCommand and then executed.
  The method SpannerCommand.GetReaderPartitions returns these.

Other details:

SpannerTransaction now has configurable behavior for releasing resources
on Dispose. It defaults to releasing back to pool, but can be changed.
It also as a readonly propery called "Shared" which marks
SpannerTransactions that are used in batch operations.
Releasing resources back into the pool is not allowed for Shared Transactions.