Add trace ids to log entries if they exist.
We do this by allowing the IHttpContextAccessor to be passed into the logger.

The only disadvantage is an IHttpContextAccessor is not avaible during the initial initialization of netcoreapp 2.0 apps so the user would have to also initialize the logger in their configure method.  The alternative would require a large change to the API of this library and how we handle DI of most of our bits.

Fixes # 1104