First prototype of BigQuery NUMERIC type support
This involves a new struct, BigQueryNumeric, which is the type
returned from BigQueryRow for NUMERIC vlaues. It has full fidelity,
along with conversions to and from System.Decimal.

Lots more work to do on performance, additional operations (e.g.
division and multiplication), lots more tests.

We may want to accept System.Decimal values for BigQueryParameter
and BigQueryInsertRow, but throw if they lose precision.