Make the AppVeyor build differ more between PR and merge
On a PR, we just run the build without coverage.
After a merge, run the build with code coverage as well.

It's easier to do this conditional work in a shell script, and we
don't really care much about the different build phases, so do it
all in a script.