Change internal representation of SpannerDbType for structs
We need to support unnamed fields and fields with duplicate names.
It makes sense to just keep a list of a new struct type with the field name (which may be empty) and type.
The type is represented as another SpannerDbType, so that we can preserve field lengths - this could be
important for EF, although it's not used directly in the wire protocol.

This change doesn't start using SpannerStruct. Tests using dictionaries as spanner structs
have been removed, to be replaced with SpannerStruct in a later commit.