Introduce WatchState for Firestore
This is the first pass at the business logic for Firestore Watch, in
terms of gathering document changes and publishing them to listeners
etc. This does *not* include any of the networking aspect, which
will be in WatchStream.

There are no unit tests other than conformance tests at the moment.

Aspects that still require further discussion (possibly after the
rest of the first implementation is ready):

- Should the callback definitely be synchronous?
- Do we definitely want the cancellation token in the callback?
- Is throwing an RpcException the correct way of signalling server
errors, or should those be more enum values?