Add a new console app to clean up test databases
The changes elsewhere in tests (in coming commits) mean we can't reliably know
when we're done with a test - and it's always possible for a machine to reboot or whatever
without cleaning up anyway. This has the additional advantage of allowing a user to look at
the database after the tests have run.