Fix the WriteTests to avoid invalid disposal
We were returning a reader with a disposed connection, which is
definitely the wrong thing to do. How we handle that situation is a
different matter, and one I'm looking into separately.

Also, one of the write tests was performing 100 writes and checking
for failures, which was left over from previous diagnostic work.