Manual fix-up of Cloud Tasks post regeneration
We need to work out what to do with the IAM methods. This commit
shows what's needed to make it compile at the moment. We may wish to
create IAM 1.1 with partial classes for the requests, at which point
this code would need fewer patch-ups. (We'd need to stop generating
the properties in *this* project.)