Improve reliability of releasing resources
This is mainly aimed at TransactionScope-based transactions, but improves other cases too.
Previously, if a connection was closed with an active SpannerDataReader, the session wouldn't be disposed properly as the client would be null.
We now retain the client we're using to create the reader, so that we can use the same one to release the session.

But more importantly, in a TransactionScope-based transaction, the connection is *intended* to be closed before we do commit/rollback etc.
We need to hand off ownership of the client to the resource manager, so that it can perform cleanup later.