More flexible URI signing
This addresses #2316 by creating a seam for the signing part of
UrlSigner.

In the future we may want to release a package specifically to make
the IAM integration simple, but for the moment we've just got a
sample in the docs.

(Another possibility would be to add a dependency on the IAM
package and include the IAM signer here, but I'd prefer not to do
that.)