Update docfx to 2.38.1
This is required to work with VS15.8, but is not sufficient: a
workaround is required, as specified in either of these issues:

- https://github.com/dotnet/docfx/issues/3225
- https://github.com/dotnet/docfx/issues/3231

The first workaround can be applied locally in the
packages/docfx.2.38.1 directory which is kept around.

This should still work with VS15.7 though, so is strictly better
than the current situation.