Start of array support in Firestore
Previous sentinel values didn't need any extra information; array removal and array union operations require the array values, so we need a different approach.

We need these in the form of a Value, but it can't look like a "normal" value - instead, we hide the array value in an extra field that isn't part of regular Value. The simplest way to do this is to use a separate proto that has the extra field, and a "kind" field that coincides with the null_value field in the original Value proto.

This is ugly, but it works :) If Google.Protobuf exposes unknown fields at some point, we can make it more efficient too.