Fix a NullReferenceException when using Precondition.None in Update
Update checks that if you specify a precondition, its proto doesn't
have Exists set to true... but in the case of Precondition.None,
there *is* no proto. As it happens, we have a handy property on
Precondition we can use anyway.

I've done a quick check for similar occurrences, and haven't found
any.