Avoid broken links in docs
Currently building a release that requires LRO or IAM will cause
broken links and odd-looking documentation. Building the packages
first fixes that, effectively by making sure that LRO/IAM
dependencies are present.

We should notice the difference next time we do a release of a
package that requires either of these.