Fixup `SubscriberClient.StopAsync` when Token is Already Cancelled
If the token passed in as the 'hard stop' token is already cancelled
then the registration will never fire and the client will peform a
soft stop instead. This can take quite some time.

The fix is to check _after_ we have registered on the token if the
token has been cancelled, and if so cancel the hard stop token source
too.