TargetedSessionPool.WaitForPoolAsync
The tests here are a little flaky as FakeScheduler doesn't behave how I expect it to.
It's reasonably clear what's expected here, but we may need to do some more work in GAX to get it non-flaky. I think it's fine to review as-is, to check that I'm testing what you think we should be testing.