Revisit active session counts yet again
This separates out "sessions the pool has asked to create" from "sessions the user has asked for": the latter can go higher than MaxActiveSessionCount (depending on configuration), but the former shouldn't.

This feels simpler than the previous attempt, which didn't actually work properly anyway.
There's *still* a potential race condition as the number of new requests is computed before we start making them (at which point we increment the counter) - but it's just as unimportant as the previous race condition, in terms of being very unlikely *and* having minimal impact.