Deflake TargetedSessionPool tests with a new FakeScheduler
This allows a lot more control over what happens when, allowing the scheduler to be run for a particular period before handing control back to the test.

Once we're agreed on the form of the new FakeScheduler, we'll port that back to GAX. I'd suggest focusing on the correctness rather than things like naming for this commit, so that we can deflake the existing tests. We can then fix naming when we move it to GAX. But I definitely want the general design and implementation validated at this point :)