Unit tests for SessionPool
These are pretty simple as TargetedSessionPool does most of the work.
The addition of file logging before/after each test is really for diagnostics, but it's so useful that I figured it's worth having it on in a more permanent way.

(Also, use the generic version of WeakReference in SessionPool itself.)