Remove a FIXME by simplifying the code
Open() makes sure we have a SessionPool, but that's all. It doesn't
need to allocate a session, therefore it doesn't matter whether or
not we have a database. We can validate quite a lot of things
without that. Admittedly if the project or instance is invalid, we
won't spot that at this point, but that's not so bad. It means we
have more consistency: an open connection always has a session pool.