Refactor serialization/deserialization code
This isn't quite a pure refactor, as ValueDeserializer is now a
static class. This does remove one potential (but previously unused)
feature: the ability to deserialize to ExpandoObject using nested
ExpandoObjects for further entities in the document. We can
introduce that later if we really want.

The benchmarks show an improvement to most cases, but not to
dictionary deserialization. That's worse than before, but not
prohibitively so. Once everything is settled, we can look at
optimization. (Basically we want to avoid reflection everywhere we
can.)

This PR is a precursor to custom serialization which will be in a
subsequent PR.