Make SigningVersion internal for now
With this in place, we can merge the V4 branch into master, allowing
us to perform other work on URL signing without creating merge
conflicts for ourselves later down the line. V4 URL signing is *not*
exposed, even if we ship a new release - the half-ready code would
be present in the assembly, but with no way to access it.