Remove legacy streaming methods and types
When Google.Cloud.Spanner.Data was first started, the code generator did not support server-side streaming methods.
Now it does, the manually-written members are redundant, so should be removed for simplicity.

We will almost certainly want to change the default timeout to 20
minutes, too - at least for queries. (For other operations, 1 minute
may be more suitable.)