Retry BatchGetDocuments RPC within GetDocumentSnapshotsAsync
This effectively treats the streaming API as if it were a unary API. We may be able to remove some of the redundancy between this and GAX in the future.
We *only* apply this to GetDocumentSnapshotsAsync - when streaming query results, we can't retry transparently as the caller can see some results before the failure.

The use of FirestoreClient.Settings doesn't play nicely with mocking. It would be nice to be able to say "You can call Settings as often as you like, but make sure no other unexpected calls are made" but I haven't been able to achieve that. It's not a big problem though.

Fixes #2760