Remove/default versions as far as possible
gRPC APIs get Grpc.Core and Google.Api.Gax.Grpc dependencies automatically. Sometimes we need more (e.g. testing) or to take a dependency from a non-gRPC project (e.g. a common project).

Version numbers for dependencies of non-GA packages can be defaulted, so we don't have to do as much to update them. Dependencies for GA packages must be specified exactly, so that if we need to create a patch we don't accidentally update the dependency.

Generated project file changes will be in a later commit.